finish SFMinionPage Refactor
